Added the ability to set knife_config_file to :auto
When the knife_config_file is set to :auto it will create a simple
knife.rb file with the chef values from the Vagrantfile. The auto
knife.rb will be created in a hidden cache folder waiting for the day
that you destroy the box. Once the box is destroyed the auto knife.rb
will be removed.